Test moved to model.local
