+++ src/main/java/egl/client/repository/local/task/TestRepository.java
@@ -1,7 +1,7 @@
 package egl.client.repository.local.task;
 
 import egl.client.model.core.task.Task;
-import egl.client.model.core.task.Test;
+import egl.client.model.local.task.Test;
 import egl.client.repository.EntityRepository;
 import org.springframework.stereotype.Repository;
