+++ src/main/java/egl/client/service/model/task/LocalTestService.java
@@ -1,15 +1,15 @@
 package egl.client.service.model.task;
 
 import egl.client.model.core.task.Task;
-import egl.client.model.core.task.Test;
-import egl.client.repository.local.task.LocalTestRepository;
+import egl.client.model.local.task.Test;
+import egl.client.repository.local.task.TestRepository;
 import egl.client.service.model.AbstractEntityService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class LocalTestService extends AbstractEntityService<Test, LocalTestRepository> {
+public class LocalTestService extends AbstractEntityService<Test, TestRepository> {
 
-    public LocalTestService(LocalTestRepository repository) {
+    public LocalTestService(TestRepository repository) {
         super(repository);
     }
